fix(graphql-connection-transformer): store graphql float as java double
Resolves: https://github.com/aws-amplify/amplify-android/issues/1031

+++ packages/amplify-codegen-appsync-model-plugin/src/scalars/index.ts
@@ -4,7 +4,7 @@ export const JAVA_SCALAR_MAP: NormalizedScalarsMap = {
   ID: 'String',
   String: 'String',
   Int: 'Integer',
-  Float: 'Float',
+  Float: 'Double',
   Boolean: 'Boolean',
   AWSDate: 'java.util.Date',
   AWSDateTime: 'java.util.Date',
